+++ hubdle-gradle-plugin/main/kotlin/repositories.accessor.kt
@@ -4,21 +4,33 @@ import org.gradle.api.artifacts.dsl.RepositoryHandler
 import org.gradle.api.artifacts.repositories.MavenArtifactRepository
 import org.gradle.kotlin.dsl.maven
 
+public const val JetBrainsComposeDevRepo: String =
+    "https://maven.pkg.jetbrains.space/public/p/compose/dev"
+
+public const val JetBrainsKotlinBootstrapRepo: String =
+    "https://maven.pkg.jetbrains.space/kotlin/p/kotlin/bootstrap"
+
+public const val JetBrainsKotlinDevRepo: String =
+    "https://maven.pkg.jetbrains.space/kotlin/p/kotlin/dev"
+
+public const val SonatypeSnapshotsRepo: String =
+    "https://oss.sonatype.org/content/repositories/snapshots"
+
 @HubdleSettingsDslMarker
 public fun RepositoryHandler.sonatypeSnapshot(action: MavenArtifactRepository.() -> Unit = {}) {
-    maven("https://oss.sonatype.org/content/repositories/snapshots", action)
+    maven(SonatypeSnapshotsRepo, action)
 }
 
 @HubdleSettingsDslMarker
 public fun RepositoryHandler.jetbrainsKotlinBootstrap(
     action: MavenArtifactRepository.() -> Unit = {}
 ) {
-    maven("https://maven.pkg.jetbrains.space/kotlin/p/kotlin/bootstrap", action)
+    maven(JetBrainsKotlinBootstrapRepo, action)
 }
 
 @HubdleSettingsDslMarker
 public fun RepositoryHandler.jetbrainsComposeDev(action: MavenArtifactRepository.() -> Unit = {}) {
-    maven("https://maven.pkg.jetbrains.space/public/p/compose/dev", action)
+    maven(JetBrainsComposeDevRepo, action)
 }
 
 @HubdleSettingsDslMarker